LayoutTests/http/tests/security/contentSecurityPolicy/script-src-none-inline-event-expected.txt

CONSOLE MESSAGE: Refused to execute inline event handler because of Content-Security-Policy.

  

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS
